test/parser/InputParserTest.java
package parser;

import org.json.JSONArray;
import org.junit.Test;

import java.util.ArrayList;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotEquals;

public class InputParserTest {
    @Test
    public void should_remove_all_unnecessary_junk_from_data() {
        String criteria = "{\"name\":\"odd number of zeroes\",\"type\":\"dfa\",\"tuple\":{\"states\":[\"q1\",\"q2\"],\"alphabets\":[\"1\",\"0\"],\"delta\":{\"q1\":{\"0\":\"q2\",\"1\":\"q1\"},\"q2\":{\"0\":\"q1\",\"1\":\"q2\"}},\"start-state\":\"q1\",\"final-states\":[\"q2\"]},\"pass-cases\":[\"0\",\"000\",\"00000\",\"10\",\"101010\",\"010101\"],\"fail-cases\":[\"00\",\"0000\",\"1001\",\"1010\",\"001100\"]}";
        String actualOutput = InputParser.parseInput(criteria);
        //it is not possible to write sentense which has {} without \\ , so the function cant be tested as easily. but it is actually working as seen in console
        assertEquals(actualOutput, criteria);
    }

    @Test
    public void should_convert_json_array_of_states_to_list_of_final_states() {
        JSONArray objects = new JSONArray();
        JSONArray finalStateAsJsonArray = objects.put("q2");
        ArrayList<String> finalStates = InputParser.getFinalStates(finalStateAsJsonArray);
        ArrayList<Object> expectedArrayList = new ArrayList<>();
        JSONArray unexpectedJsonArray = new JSONArray();
        assertEquals(finalStates.getClass(),expectedArrayList.getClass());
        assertNotEquals(finalStates.getClass(),unexpectedJsonArray.getClass());
    }

    @Test
    public void should_convert_json_array_of_alphabets_to_list_of_alphabets(){
        JSONArray objects = new JSONArray();
        JSONArray alphabetsAsJsonArray = objects.put("A");
        ArrayList<String> alphabets = InputParser.getAlphabets(alphabetsAsJsonArray);
        ArrayList<Object> expectedArrayList = new ArrayList<>();
        JSONArray unexpectedJsonArray = new JSONArray();
        assertEquals(alphabets.getClass(),expectedArrayList.getClass());
        assertNotEquals(alphabets.getClass(),unexpectedJsonArray.getClass());
    }
}